Code Cleanup -- Optional

+++ src/main/java/biomesoplenty/common/fluids/blocks/BlockBloodFluid.java
@@ -73,7 +73,7 @@ public class BlockBloodFluid extends BlockFluidClassic
 
         if (flag)
         {
-            Integer integer = (Integer)state.getValue(LEVEL);
+            Integer integer = state.getValue(LEVEL);
 
             if (integer.intValue() == 0)
             {
