+++ src/main/java/biomesoplenty/common/fluids/blocks/BlockQuicksandFluid.java
@@ -78,7 +78,7 @@ public class BlockQuicksandFluid extends BlockFluidClassic
 
         if (flag)
         {
-            Integer integer = (Integer)state.getValue(LEVEL);
+            Integer integer = state.getValue(LEVEL);
 
             if (integer.intValue() == 0)
             {
